requirements.txt
pyopenssl>=24.1.0
paramiko>=3.0.0
imcsdk>=0.9.13
jsonschema>=3.2.0
python-docx>=1.0.1
requests>=2.23.0
ucsmsdk>=0.9.16
ucscsdk>=0.9.0.4
matplotlib>=3.2.1
natsort>=7.1.1
networkx[default]>=2.4
Pillow>=10.0.0
netaddr>=0.7.19
pyyaml>=5.3.1
packaging>=20.4
Flask>=2.0.0
Flask-Cors>=3.0.10
cryptography>=42.0.2
urllib3>=1.25.9
intersight>=1.0.11.15473
SQLAlchemy>=2.0.23
werkzeug>=1.0.1
numpy>=1.19.4
tabulate>=0.8.9
reportlab>=3.6.9
scp>=0.14.4

# EASYUCS-620: Restricting cffi version to 1.15.0. This is a workaround to avoid the error "Invalid DER encoding inside the PEM file"
# TODO: To root cause the issue, we need to do more debugging inside cffi code.
cffi==1.15.0

# NXOS dependencies
python-dotenv
